refactor(race): RaceManager 내 테스트 불필요한 로직 Race로 이동
